requirements.txt
asgiref==3.7.2
Authlib==1.2.1
backports.zoneinfo==0.2.1
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==41.0.5
defusedxml==0.7.1
dj-rest-auth==5.0.1
django-allauth==0.58.2
Django==4.2.7
djangorestframework-simplejwt==5.3.0
djangorestframework==3.14.0
idna==3.4
oauthlib==3.2.2
pip==23.3.1
psycopg2-binary==2.9.9
pycparser==2.21
PyJWT==2.8.0
python-dateutil==2.8.2
python-decouple==3.8
python3-openid==3.2.0
pytz==2023.3.post1
requests-oauthlib==1.3.1
requests==2.31.0
setuptools==68.2.2
six==1.16.0
sqlparse==0.4.4
typing-extensions==4.8.0
tzdata==2023.3
urllib3==2.0.7
wheel==0.41.3
xero-python==1.25.0
